walter/requirements.txt
Django==2.2.2
django-debug-toolbar==1.11
django-filter==2.1.0
django-redis==4.10.0
djangorestframework==3.9.4
Markdown==3.1.1
psycopg2==2.8.3
pytz==2019.1
redis==3.2.1
sqlparse==0.3.0
